consolidated user and room models
