+++ internal/client/http/rooms.go
@@ -8,6 +8,7 @@ import (
 	"io"
 	"net/http"
 
+	"github.com/Abi-Liu/TextTunnel/internal/models"
 	"github.com/google/uuid"
 	"nhooyr.io/websocket"
 )
@@ -26,7 +27,7 @@ func (c *HttpClient) ConnectToSocket(id uuid.UUID) (*websocket.Conn, context.Con
 	return conn, ctx, cancel, nil
 }
 
-func (c *HttpClient) FetchRooms() ([]Room, error) {
+func (c *HttpClient) FetchRooms() ([]models.Room, error) {
 	res, err := c.Get(BASE_URL + "/rooms")
 	if err != nil {
 		return nil, err
@@ -37,7 +38,7 @@ func (c *HttpClient) FetchRooms() ([]Room, error) {
 		return nil, err
 	}
 
-	rooms := []Room{}
+	rooms := []models.Room{}
 	err = json.Unmarshal(data, &rooms)
 	if err != nil {
 		return nil, err
@@ -46,30 +47,30 @@ func (c *HttpClient) FetchRooms() ([]Room, error) {
 	return rooms, nil
 }
 
-func (c *HttpClient) CreateRoom(name string) (Room, error) {
+func (c *HttpClient) CreateRoom(name string) (models.Room, error) {
 	type Req struct {
 		Name string `json:"name"`
 	}
 	params := Req{Name: name}
 	data, err := json.Marshal(params)
 	if err != nil {
-		return Room{}, fmt.Errorf("failed to Marshal to json: %s", err)
+		return models.Room{}, fmt.Errorf("failed to Marshal to json: %s", err)
 	}
 	reader := bytes.NewReader(data)
 	res, err := c.Post(BASE_URL+"/rooms", reader)
 	if err != nil {
-		return Room{}, fmt.Errorf("failed to create room: %s", err)
+		return models.Room{}, fmt.Errorf("failed to create room: %s", err)
 	}
 	defer res.Body.Close()
 
 	dat, err := io.ReadAll(res.Body)
 	if err != nil {
-		return Room{}, fmt.Errorf("failed to read response body: %s", err)
+		return models.Room{}, fmt.Errorf("failed to read response body: %s", err)
 	}
-	room := Room{}
+	room := models.Room{}
 	err = json.Unmarshal(dat, &room)
 	if err != nil {
-		return Room{}, fmt.Errorf("failed to Unmarshal data: %s", err)
+		return models.Room{}, fmt.Errorf("failed to Unmarshal data: %s", err)
 	}
 
 	return room, nil
