--- internal/client/http/types.go
@@ -1,28 +1,0 @@
-package http
-
-import (
-	"time"
-
-	"github.com/google/uuid"
-)
-
-type User struct {
-	ID        uuid.UUID `json:"id"`
-	Username  string    `json:"username"`
-	CreatedAt time.Time `json:"created_at"`
-	UpdatedAt time.Time `json:"updated_at"`
-	ApiKey    string    `json:"api_key"`
-}
-
-type Room struct {
-	ID        uuid.UUID `json:"id"`
-	Name      string    `json:"name"`
-	CreatedAt time.Time `json:"created_at"`
-	UpdatedAt time.Time `json:"updated_at"`
-	OwnerId   uuid.UUID `json:"owner_id"`
-	CreatorId uuid.UUID `json:"creator_id"`
-}
-
-type Error struct {
-	Error string `json:"error"`
-}
