+++ internal/models/error.go
@@ -1,0 +1,5 @@
+package models
+
+type Error struct {
+	Error string `json:"error"`
+}
