+++ internal/models/users.go
@@ -1,0 +1,36 @@
+package models
+
+import (
+	"time"
+
+	"github.com/Abi-Liu/TextTunnel/internal/database"
+	"github.com/google/uuid"
+)
+
+type User struct {
+	ID        uuid.UUID `json:"id"`
+	Username  string    `json:"username"`
+	CreatedAt time.Time `json:"created_at"`
+	UpdatedAt time.Time `json:"updated_at"`
+	ApiKey    string    `json:"api_key"`
+}
+
+func DatabaseUserToUser(user database.User) User {
+	return User{
+		ID:        user.ID,
+		Username:  user.Username,
+		CreatedAt: user.CreatedAt,
+		UpdatedAt: user.UpdatedAt,
+		ApiKey:    user.ApiKey,
+	}
+}
+
+func DatabaseUsersToUsers(users []database.User) []User {
+	res := make([]User, len(users))
+
+	for i, user := range users {
+		res[i] = DatabaseUserToUser(user)
+	}
+
+	return res
+}
