+++ internal/server/user_handler.go
@@ -8,8 +8,8 @@ import (
 	"strings"
 
 	"github.com/Abi-Liu/TextTunnel/internal/database"
+	"github.com/Abi-Liu/TextTunnel/internal/models"
 	"github.com/Abi-Liu/TextTunnel/internal/server/auth"
-	"github.com/Abi-Liu/TextTunnel/internal/server/models"
 	"github.com/google/uuid"
 )
